fix: implement code audit findings - security, correctness, DRY cleanup
- AC1: Defense-in-depth path traversal fix in _get_image_dir
- AC2: Corrupt image handling with try/except in _load_image
- AC3: KeyError guard for GRID_DIMENSIONS lookup
- AC4: Replace scipy.stats.mode with np.bincount (drop scipy dep)
- AC5: Fix quantization precision with .round() before uint8 cast
- AC6: Standardize ColorPalette.colors_rgb to uint8 everywhere
- AC7: Add startup env var validation with clear error messages
- AC8: Add CORSMiddleware with configurable allowed origins
- AC9: Add Content-Security-Policy header middleware
- AC10: Define typed PaletteEntry Pydantic model for API contracts
- AC11: Centralize test helpers in conftest.py (DRY)
- AC12: Extract shared perceived_brightness utility
- AC13: Pin dependency upper bounds in requirements.txt
- AC14: Separate test deps into requirements-dev.txt
- AC15-18: Remove dead ErrorResponse, fix comments, batch LAB conversion

+++ tests/test_palette_edit.py
@@ -9,7 +9,7 @@ def _make_palette(n: int = 5) -> ColorPalette:
     """Create a palette with n distinct colors."""
     rng = np.random.default_rng(42)
     colors = rng.integers(0, 255, size=(n, 3), dtype=np.uint8)
-    return ColorPalette(colors_rgb=colors.astype(np.float64))
+    return ColorPalette(colors_rgb=colors)
 
 
 def _make_sheet(palette: ColorPalette, rows: int = 4, cols: int = 4) -> MosaicSheet:
@@ -39,7 +39,7 @@ class TestColorSwapUpdatesPalette:
         """Updating colors_rgb[i] changes the palette color at that index."""
         palette = _make_palette(5)
         old_color = palette.colors_rgb[2].copy()
-        new_rgb = np.array([255, 0, 128], dtype=np.float64)
+        new_rgb = np.array([255, 0, 128], dtype=np.uint8)
 
         palette.colors_rgb[2] = new_rgb
 
@@ -99,7 +99,7 @@ class TestSimilarColorWarning:
                     [0, 0, 255],  # index 2: blue
                     [250, 5, 0],  # index 3: very similar to red (index 0)
                 ],
-                dtype=np.float64,
+                dtype=np.uint8,
             )
         )
         # Editing index 3 to a color very similar to index 0
@@ -120,7 +120,7 @@ class TestSimilarColorWarning:
                     [0, 255, 0],
                     [0, 0, 255],
                 ],
-                dtype=np.float64,
+                dtype=np.uint8,
             )
         )
         warnings = _compute_palette_warnings(palette, color_index=0)
@@ -141,7 +141,7 @@ class TestDuplicateColorWarning:
                     [0, 0, 0],
                     [100, 200, 50],  # duplicate of index 0
                 ],
-                dtype=np.float64,
+                dtype=np.uint8,
             )
         )
         warnings = _compute_palette_warnings(palette, color_index=2)
